refactor(mason-*): Separating `mason-*` from `lspconfig` and `null-ls`
